Remove eth outcome/allocation helpers from index
