+++ packages/wallet-specs/src/protocols/conclude-channel/protocol.ts
@@ -2,17 +2,11 @@ import { Machine } from 'xstate';
 
 import * as VirtualDefundingAsHub from '../virtual-defunding-as-hub/protocol';
 import * as VirtualDefundingAsLeaf from '../virtual-defunding-as-leaf/protocol';
-import {
-  MachineFactory,
-  FINAL,
-  checkThat,
-  getEthAllocation,
-  ethAllocationOutcome,
-  add,
-} from '../..';
+import { MachineFactory, FINAL, checkThat, add } from '../..';
 import { IStore } from '../../store';
 import { getDetaAndInvoke } from '../../machine-utils';
 import { isIndirectFunding } from '../../ChannelStoreEntry';
+import { getEthAllocation, ethAllocationOutcome } from '../../calculations';
 
 import { SupportState } from '..';
