+++ packages/wallet-specs/src/protocols/direct-funding/protocol.ts
@@ -5,16 +5,9 @@ import { bigNumberify } from 'ethers/utils';
 import { HashZero, AddressZero } from 'ethers/constants';
 
 import { getDetaAndInvoke } from '../../machine-utils';
-import {
-  add,
-  subtract,
-  ethAllocationOutcome,
-  getEthAllocation,
-  FINAL,
-  MachineFactory,
-  gt,
-} from '../../';
+import { add, subtract, FINAL, MachineFactory, gt } from '../../';
 import { IStore } from '../../store';
+import { getEthAllocation, ethAllocationOutcome } from '../../calculations';
 
 import { SupportState, Depositing } from '..';
