+++ packages/wallet-specs/src/protocols/virtual-defunding-as-leaf/protocol.ts
@@ -1,10 +1,11 @@
 import { assign } from 'xstate';
-import { isAllocationOutcome, Allocation } from '@statechannels/nitro-protocol';
+import { Allocation } from '@statechannels/nitro-protocol';
 
-import { Without, ethAllocationOutcome, checkThat, getEthAllocation } from '../../';
+import { Without, checkThat } from '../../';
 import { isIndirectFunding, isVirtualFunding } from '../../ChannelStoreEntry';
 import { store } from '../../temp-store';
 import * as LedgerUpdate from '../ledger-update/protocol';
+import { getEthAllocation, ethAllocationOutcome } from '../../calculations';
 const PROTOCOL = 'virtual-defunding-as-leaf';
 
 export interface Init {
